Fix bug #2
This change handles the bug where the tool was being triggered even when we didn't have any change or when the change was outside a IFFT block. Now the tool is triggered only when a change is made in the repository and this change is placed inside a IFFT block.